separate package into core, lifi and uniswap
