--- src/abis/index.ts
@@ -1,3 +1,0 @@
-import ERC20 from './ERC20.json';
-
-export { ERC20 };
