--- src/core/hooks.tsx
@@ -1,42 +1,0 @@
-import { useCallback, useState } from 'react';
-import { TokenInfo } from '../tokens';
-import { isMetaMaskProvider, uriToHttp } from './utils';
-import { JsonRpcProvider } from '@ethersproject/providers';
-
-export const useAddTokenToMetamask = (
-  token: TokenInfo,
-  library: JsonRpcProvider
-): {
-  addToken: () => void;
-  success: boolean | null;
-} => {
-  const [success, setSuccess] = useState<boolean | null>(null);
-
-  const addToken = useCallback(() => {
-    if (library && isMetaMaskProvider(library) && token) {
-      const { address, symbol, decimals, logoURI } = token;
-      const logoURL = logoURI ? uriToHttp(logoURI)[0] : '';
-      library.provider
-        .request({
-          method: 'wallet_watchAsset',
-          params: {
-            type: 'ERC20',
-            options: {
-              address: address,
-              symbol: symbol,
-              decimals: decimals,
-              image: logoURL,
-            },
-          },
-        })
-        .then(success => {
-          setSuccess(success);
-        })
-        .catch(() => setSuccess(false));
-    } else {
-      setSuccess(false);
-    }
-  }, [library, token]);
-
-  return { addToken, success };
-};
