+++ src/lifi/index.ts
@@ -1,0 +1,1 @@
+export const lifi = 'lifi';
